+++ test/nn/conv/test_eg_conv.py
@@ -53,7 +53,7 @@ def test_eg_conv(aggregators, add_self_loops):
         assert torch.allclose(conv(x, adj2.t()), out, atol=1e-2)
 
     if is_full_test():
-        jit = torch.jit.script(conv.jittable())
+        jit = torch.jit.script(conv)
         assert torch.allclose(jit(x, edge_index), out, atol=1e-2)
 
         if torch_geometric.typing.WITH_TORCH_SPARSE: