+++ test/nn/conv/test_pna_conv.py
@@ -33,7 +33,7 @@ def test_pna_conv(divide_input):
         assert torch.allclose(conv(x, adj.t()), out, atol=1e-6)
 
     if is_full_test():
-        jit = torch.jit.script(conv.jittable())
+        jit = torch.jit.script(conv)
         assert torch.allclose(jit(x, edge_index, value), out, atol=1e-6)
 
         if torch_geometric.typing.WITH_TORCH_SPARSE: