+++ torch_geometric/datasets/upfd.py
@@ -9,7 +9,7 @@ import torch
 from torch_geometric.data import (
     Data,
     InMemoryDataset,
-    download_url,
+    download_google_url,
     extract_zip,
 )
 from torch_geometric.io import read_txt_array
@@ -75,9 +75,7 @@ class UPFD(InMemoryDataset):
         force_reload (bool, optional): Whether to re-process the dataset.
             (default: :obj:`False`)
     """
-    url = 'https://docs.google.com/uc?export=download&id={}&confirm=t'
-
-    ids = {
+    file_ids = {
         'politifact': '1KOmSrlGcC50PjkvRVbyb_WoWHVql06J-',
         'gossipcop': '1VskhAQ92PrT4sWEKQ2v2-AJhEcpp4A81',
     }
@@ -93,13 +91,16 @@ class UPFD(InMemoryDataset):
         pre_filter: Optional[Callable] = None,
         force_reload: bool = False,
     ) -> None:
+        assert name in ['politifact', 'gossipcop']
+        assert split in ['train', 'val', 'test']
+
         self.root = root
         self.name = name
         self.feature = feature
+
         super().__init__(root, transform, pre_transform, pre_filter,
                          force_reload=force_reload)
 
-        assert split in ['train', 'val', 'test']
         path = self.processed_paths[['train', 'val', 'test'].index(split)]
         self.load(path)
 
@@ -123,7 +124,8 @@ class UPFD(InMemoryDataset):
         return ['train.pt', 'val.pt', 'test.pt']
 
     def download(self) -> None:
-        path = download_url(self.url.format(self.ids[self.name]), self.raw_dir)
+        id = self.file_ids[self.name]
+        path = download_google_url(id, self.raw_dir, 'data.zip')
         extract_zip(path, self.raw_dir)
         os.remove(path)